add frontend-test, unescape ParseString, jinja template
* Added generic tree walk to the stream API, allows arbitrary navigation
  of json-serializable structures
* Add python script to code-generate the stream API rather than using C
  macros.
* Add utilities to escape/unescape strings for JSON serialization.
* Fix missing backslash in regex for STRING_LITERAL token
* Cleanup some compiler warnings
* Add a frontend test to execute the demo program on some canonical
  input and ensure that the lex/parse/markup output matches expected
  outputs
* ParseString will now unescape the contents
* Moved parse/emit functions to their own files
* Moved parse/emit functions out of the stream namespace
* Merge _tpl.h files into -> .h
* Make json_gen use a jinja template
* Replace remaining printf() with LOG() for parse errors

+++ json/test/parser_test.cc
@@ -1,11 +1,12 @@
 // Copyright 2018 Josh Bialkowski <josh.bialkowski@gmail.com>
 #include <array>
+
 #include <gtest/gtest.h>
 #include <re2/re2.h>
 
 #include "json/item.h"
 
-using namespace json;
+using namespace json;  // NOLINT
 
 TEST(ParserTest, Re2CanParseJSONNumbers) {
   int64_t integer;
@@ -71,7 +72,7 @@ TEST(ItemParserTest, TestKnownParsings) {
 
   Error parse_error{};
   item::ItemParser parser(g_item_store_.begin(), g_item_store_.end());
-  for (size_t idx = 0; idx < ntokens; ++idx) {
+  for (size_t idx = 0; idx < static_cast<size_t>(ntokens); ++idx) {
     ASSERT_EQ(0, parser.Consume(g_token_store_[idx], &parse_error))
         << "Error(" << parse_error.code << "): " << parse_error.msg
         << " for token " << idx;
